Plugin Check, WP 6.4

+++ admin.php
@@ -20,8 +20,8 @@ add_action('admin_menu', 'leafext_dsgvo_add_page', 100);
 function leafext_dsgvo_init(){
 	// Create Setting
 	$section_group = 'leafext_dsgvo';
- 	$section_name = 'leafext_dsgvo';
- 	$validate = 'leafext_validate_dsgvo';
+	$section_name = 'leafext_dsgvo';
+	$validate = 'leafext_validate_dsgvo';
 	register_setting( $section_group, $section_name, $validate );
 
 	// Create section of Page
@@ -80,16 +80,17 @@ add_action( 'admin_init', 'leafext_dsgvo_init' );
 function leafext_dsgvo_form() {
 	$setting = leafext_okay();
 	echo '<textarea name="leafext_dsgvo[text]" type="textarea" cols="80" rows="5">';
-	echo $setting;
+	echo esc_textarea($setting);
 	echo '</textarea>';
 }
 
 function leafext_dsgvo_form_mapurl() {
 	$image="";
-  $options = get_option( 'leafext_dsgvo' );
-  if ( is_array ($options) && $options['mapurl'] != "" ) $image=$options['mapurl'];
+	$options = get_option( 'leafext_dsgvo' );
+	if ( is_array ($options) && $options['mapurl'] != "" ) $image=$options['mapurl'];
 	$placeholder = ($image == "") ? LEAFEXT_DSGVO_PLUGIN_URL.'/map.png' : $image;
-	echo '<input type="url" size="80" placeholder="'.$placeholder.'" name="leafext_dsgvo[mapurl]" value="'.$image.'" /><p>'.__('URL to Background Image','extensions-leaflet-map-dsgvo').'</p>';
+	echo '<input type="url" size="80" placeholder="'.esc_textarea($placeholder).'" name="leafext_dsgvo[mapurl]" value="'.esc_url($image).
+	'" /><p>'.__('URL to Background Image','extensions-leaflet-map-dsgvo').'</p>';
 }
 
 function leafext_dsgvo_form_cookie() {
@@ -101,11 +102,11 @@ function leafext_dsgvo_form_cookie() {
 		$cookie = "365";
 		$form = 'placeholder="'.$cookie.'"';
 	}
-	echo '<input type="number" size="3" max="365" name="leafext_dsgvo[cookie]" '.$form.'"> '.__('days','extensions-leaflet-map-dsgvo');
+	echo '<input type="number" size="5" min="1" max="365" name="leafext_dsgvo[cookie]" '.$form.'"> '.__('days','extensions-leaflet-map-dsgvo');
 }
 
 function leafext_dsgvo_form_count() {
-  $options = get_option( 'leafext_dsgvo' );
+	$options = get_option( 'leafext_dsgvo' );
 	$count = ( is_array ($options) && isset($options['count']) ) ? filter_var($options['count'],FILTER_VALIDATE_BOOLEAN) : false;
 	echo '<input type="radio" name="leafext_dsgvo[count]" value="1" ';
 	echo $count ? 'checked' : '' ;
@@ -124,6 +125,7 @@ function leafext_dsgvo_form_okay() {
 
 // Sanitize and validate input. Accepts an array, return a sanitized array.
 function leafext_validate_dsgvo($options) {
+	check_admin_referer('leafext_dsgvo', 'leafext_dsgvo_nonce');
 	if (isset($_POST['submit'])) {
 		//var_dump($options); wp_die();
 		if ($options['cookie'] == "0" || $options['cookie'] == "" )	$options['cookie'] = "365";
@@ -139,32 +141,43 @@ function leafext_validate_dsgvo($options) {
 
 // Erklaerung / Hilfe
 function leafext_dsgvo_help() {
-	$text = file_get_contents( LEAFEXT_DSGVO_PLUGIN_DIR . "/readme.md" );
-	//[Extensions for Leaflet Map](https://de.wordpress.org/plugins/extensions-leaflet-map/)
-	$suchmuster = array(
-		'/\[(.+)\]\((.+)\)/i',
-		'/(### )(.*)/',
-		'/(## )(.*)/',
-		'/(# )(.*)/',
-		'/  /',
-	);
-	$ersetzung = array(
-		'<a href="${2}">${1}</a>',
-		'<h3>${2}</h3>',
-		'<h2>${2}</h2>',
-		'<h1>${2}</h1>',
-		'<br>',
-	);
- 	$text = preg_replace($suchmuster, $ersetzung, $text);
-	echo '<div style="width:80%">'.$text.'</div>';
-	echo '<h3>'.__('Settings','extensions-leaflet-map-dsgvo').'</h3>';
-	echo '<p>'.__('Test it in a private browser window.','extensions-leaflet-map-dsgvo');
+	// Call globals
+	global $wp_filesystem;
+	// Initiate
+	WP_Filesystem();
+	$local_file = LEAFEXT_DSGVO_PLUGIN_DIR . "/readme.md";
+	$text = '';
+	if ( $wp_filesystem->exists( $local_file ) ) {
+		$text = $wp_filesystem->get_contents( $local_file ) ;
+		//[Extensions for Leaflet Map](https://de.wordpress.org/plugins/extensions-leaflet-map/)
+		$suchmuster = array(
+			'/\[(.+)\]\((.+)\)/i',
+			'/(### )(.*)/',
+			'/(## )(.*)/',
+			'/(# )(.*)/',
+			'/  /',
+		);
+		$ersetzung = array(
+			'<a href="${2}">${1}</a>',
+			'<h3>${2}</h3>',
+			'<h2>${2}</h2>',
+			'<h1>${2}</h1>',
+			'<br>',
+		);
+		$text = preg_replace($suchmuster, $ersetzung, $text);
+		echo '<div style="width:80%">'.$text.'</div>';
+		echo '<h3>'.__('Settings','extensions-leaflet-map-dsgvo').'</h3>';
+		echo '<p>'.__('Test it in a private browser window.','extensions-leaflet-map-dsgvo');
+	} else {
+		echo "Error";
+	}
 }
 
 // Draw the menu page itself
 function leafext_dsgvo_do_page (){
 	echo '<form method="post" action="options.php">';
 	settings_fields('leafext_dsgvo');
+	wp_nonce_field( 'leafext_dsgvo', 'leafext_dsgvo_nonce' );
 	do_settings_sections( 'leafext_dsgvo' );
 	submit_button();
 	submit_button( __( 'Reset', 'extensions-leaflet-map-dsgvo' ), 'delete', 'delete', false);
